requirements.txt
fastapi
uvicorn
pandas
locust
google-genai
python-dotenv
requests